requirements.txt
google-genai
python-dotenv
sqlalchemy
pymysql
langchain
langchain-community
langchain-google-genai
streamlit
